Moved tests from the app module to the relevant httpsupgrade modules. Converted some android tests to unit tests.
