+++ httpsupgrade/httpsupgrade-impl/build.gradle
@@ -43,15 +43,14 @@ dependencies {
     implementation Square.retrofit2.retrofit
     implementation Square.retrofit2.converter.moshi
 
+   // Testing dependencies
     testImplementation project(':common-test')
     testImplementation Testing.junit4
-    androidTestImplementation Testing.junit4
-    testImplementation AndroidX.archCore.testing
-    androidTestImplementation AndroidX.archCore.testing
-    testImplementation AndroidX.test.ext.junit
+    androidTestUtil AndroidX.test.orchestrator
+    androidTestImplementation AndroidX.test.runner
+    androidTestImplementation "org.mockito:mockito-android:_"
+    androidTestImplementation "org.mockito.kotlin:mockito-kotlin:_"
     androidTestImplementation AndroidX.test.ext.junit
-    testImplementation "org.mockito.kotlin:mockito-kotlin:_"
-    testImplementation Testing.robolectric
 
     // WorkManager
     implementation AndroidX.work.runtimeKtx
