+++ httpsupgrade/httpsupgrade-impl/src/test/java/com/duckduckgo/httpsupgrade/impl/HttpsBloomFilterSpecJsonTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018 DuckDuckGo
+ * Copyright (c) 2023 DuckDuckGo
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.app.httpsupgrade.api
+package com.duckduckgo.httpsupgrade.impl
 
 import com.duckduckgo.httpsupgrade.store.HttpsBloomFilterSpec
 import com.squareup.moshi.Moshi
