+++ httpsupgrade/httpsupgrade-impl/src/test/java/com/duckduckgo/httpsupgrade/impl/HttpsFalsePositivesJsonTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018 DuckDuckGo
+ * Copyright (c) 2023 DuckDuckGo
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,8 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.app.httpsupgrade.api
+package com.duckduckgo.httpsupgrade.impl
 
-import com.duckduckgo.httpsupgrade.impl.HttpsFalsePositivesJsonAdapter
 import com.duckduckgo.httpsupgrade.store.HttpsFalsePositiveDomain
 import com.squareup.moshi.JsonAdapter
 import com.squareup.moshi.Moshi
