+++ httpsupgrade/httpsupgrade-store/src/test/java/com/duckduckgo/httpsupgrade/store/HttpsBloomFilterSpecDaoTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018 DuckDuckGo
+ * Copyright (c) 2023 DuckDuckGo
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,26 +14,41 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.app.httpsupgrade.store
+package com.duckduckgo.httpsupgrade.store
 
+import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.room.Room
+import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
-import com.duckduckgo.httpsupgrade.store.HttpsBloomFilterSpec
-import com.duckduckgo.httpsupgrade.store.HttpsBloomFilterSpecDao
-import com.duckduckgo.httpsupgrade.store.HttpsUpgradeDatabase
+import com.duckduckgo.app.CoroutineTestRule
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.test.runTest
 import org.junit.After
 import org.junit.Assert.*
 import org.junit.Before
+import org.junit.Rule
 import org.junit.Test
+import org.junit.runner.RunWith
 
+@ExperimentalCoroutinesApi
+@RunWith(AndroidJUnit4::class)
 class HttpsBloomFilterSpecDaoTest {
+
+    @get:Rule
+    @Suppress("unused")
+    var instantTaskExecutorRule = InstantTaskExecutorRule()
+
+    @get:Rule
+    var coroutinesTestRule = CoroutineTestRule()
 
     private lateinit var db: HttpsUpgradeDatabase
     private lateinit var dao: HttpsBloomFilterSpecDao
 
     @Before
     fun before() {
-        db = Room.inMemoryDatabaseBuilder(InstrumentationRegistry.getInstrumentation().targetContext, HttpsUpgradeDatabase::class.java).build()
+        db = Room.inMemoryDatabaseBuilder(InstrumentationRegistry.getInstrumentation().targetContext, HttpsUpgradeDatabase::class.java)
+            .allowMainThreadQueries()
+            .build()
         dao = db.httpsBloomFilterSpecDao()
     }
 
@@ -43,25 +58,25 @@ class HttpsBloomFilterSpecDaoTest {
     }
 
     @Test
-    fun whenModelIsEmptyThenGetIsNull() {
+    fun whenModelIsEmptyThenGetIsNull() = runTest {
         assertNull(dao.get())
     }
 
     @Test
-    fun whenModelIsInsertedThenGetIsNotNull() {
+    fun whenModelIsInsertedThenGetIsNotNull() = runTest {
         dao.insert(HttpsBloomFilterSpec(errorRate = 0.1, bitCount = 1000, totalEntries = 55, sha256 = "abc"))
         assertNotNull(dao.get())
     }
 
     @Test
-    fun whenNewModelIsInsertedThenGetIsNotNullAndDetailsUpdates() {
+    fun whenNewModelIsInsertedThenGetIsNotNullAndDetailsUpdates() = runTest {
         dao.insert(HttpsBloomFilterSpec(bitCount = 1000, errorRate = 0.1, totalEntries = 55, sha256 = "abc"))
         dao.insert(HttpsBloomFilterSpec(bitCount = 2000, errorRate = 0.2, totalEntries = 60, sha256 = "xyz"))
 
         val specification = dao.get()
         assertNotNull(specification)
         assertEquals(2000, specification!!.bitCount)
-        assertEquals(0.2, specification!!.errorRate, 0.01)
+        assertEquals(0.2, specification.errorRate, 0.01)
         assertEquals(60, specification.totalEntries)
         assertEquals("xyz", specification.sha256)
     }
